Extract post-turn PR transition handling
